tools/{fru-generator,libipmi}: copied from PTS::1564157
The contents of this commit are part of the "fmcadc100m14b4cha" branch
of the "pts" project (Python Test Suite) on ohwr, commit 1564157 (Nov
13 2012).

Files are renamed like this:

 test/fmceeprom/python/fmc_eeprom_example.py  -> tools/fru-generator
 test/fmceeprom/python/libipmi/               -> tools/libipmi/
 test/fmceeprom/python/fmc_eeprom.py          -> tools/libipmi/fmc_eeprom.py

The rationale is that tools/fru-generator will be called by our users,
while fmc_eeprom.py is part of the library.

Later commits in fmc-bus change the files (now copied literally) to fit
the different use case.

Matthieu is the author of this code with Manohar Vanga, but neither is
reposnsible for this commit. The committer marked Matthieu as author
so "git blame" will show him not me.

It is lifted here (and renamed) unchanged before being customized.