fru-generator: several changes, to support user configuration
This addes command line support and environment support to specify the
values to be set in the EEPROM. Removes the dumping function
(tools/fru-dump will soon be there), removes generation of the fake
internal use area and supports the soon-to-be-documented defaults.

It is still missing a lot (i.e., power consumption is still hardwired)
but it can be used for the initial work on eeprom images.

Again, thanks to Manohar Vanga and Matthieu Cattin for libipmi and
this tools.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>