kernel/fmc-fakedev: bugfix: allocate dynamically the device
This changes a while how stuff happens in fmc-fakedev, according to
comments by Emilio Cota who noted how I got it badly hacked and not
really stable. This now mallocs and uses dev->release as it should.

It also fills the new fields of fmc_device that are mandated by the
current version of the core (carrier_data, nr_slots, eeprom_addr).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>