fmc-core: bug fix requiring incompatible changes in registration
My previous code for device management was wrong, because I didn't
understand the allocation. This fixes the thing, by properly using the
release function.

This change however requires that all fmc devices are allocated
separately (not as an array) and are not freed by the registering
entity.  So this introduces an incompatible change in the prototype
for register_n and unregister_n -- which is noted at build time, but
also an incompatible change in how a device must be allocated/freed
by the carrier that registers a single device. This change is
not introducing a build error, but I hope the only external users by now
are the SPEC and the SVEC.

fmc-fakedev is fixed by this same commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>